+++ qualtran/bloqs/arithmetic/trigonometric/arctan.py
@@ -1,4 +1,16 @@
-#  Copyright 2023 Google LLC
+#  Copyright 2024 Google LLC
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -32,6 +44,8 @@ class ArcTan(GateWithRegisters, cirq.ArithmeticGate):  # type: ignore[misc]
         target_bitsize: The bitsize of output register. The computed quantity,
             $\abs(-2 * \arctan(x) / \pi)$ is stored as a fixed-length binary approximation
             in the output register of size `target_bitsize`.
+
+    TODO missing reference
     """
 
     selection_bitsize: int
