+++ qualtran/bloqs/arithmetic/trigonometric/arctan_test.py
@@ -1,4 +1,16 @@
-#  Copyright 2023 Google LLC
+#  Copyright 2024 Google LLC
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -17,7 +29,7 @@ import numpy as np
 import pytest
 
 from qualtran import QFxp
-from qualtran.bloqs.mean_estimation.arctan import ArcTan
+from qualtran.bloqs.arithmetic.trigonometric.arctan import ArcTan
 from qualtran.cirq_interop.t_complexity_protocol import t_complexity, TComplexity
